requirements.txt
sqlalchemy<2.0
Flask~=2.2.2
PyJWT~=2.6.0